added blog page for models and schemas.
