+++ Models/blog.js
@@ -31,39 +31,4 @@ const blogPostSchema = new mongoose.Schema(
 
          module.exports = BlogPost;
 
-         const mongoose = require('mongoose');
 
-// BlogPage schema to represent an entire page of blog posts and related details
-const blogPageSchema = new mongoose.Schema({
-    pageTitle: {
-        type: String,
-        required: true,
-    },
-    posts: [{
-        type: mongoose.Schema.Types.ObjectId, 
-        ref: 'BlogPost', // Reference to blog posts
-    }],
-    author: {
-        type: String, // Can also be changed to reference an 'Author' schema if needed
-        required: true,
-    },
-    created: {
-        type: Date,
-        default: Date.now,
-    },
-    updated: {
-        type: Date,
-        default: Date.now,
-    },
-    description: {
-        type: String, // A description for the blog page
-    },
-    bannerImage: {
-        type: String, // URL or path to a banner image for the blog page
-    }
-});
-
-// Create a model for BlogPage
-const BlogPage = mongoose.model('BlogPage', blogPageSchema);
-
-module.exports = BlogPage;
