+++ app.js
@@ -7,25 +7,21 @@ const { dbConnect } = require("./db");
 const PORT = process.env.PORT;
 
 // Import routers
-const StoryRoute = require('./routers/storyRoute');
+const storyRoute = require('./routers/storyRoute');
 const newsletterRoutes = require('./routers/newsletterRoute'); 
-//const blogRoutes = require('./routers/blogRoutes');
+const blogRoutes = require('./routers/blogroutes');
 
-
-// const blogRoutes = require('./routers/blogRoutes')
 //Middleware
 app.use(express.json());
 app.use(cors());
 app.use(cookieParser());
-// app.use("/blog", blogRoutes);
 
 // use routes
-//app.use("/newsletter", newsletterRoutes); 
-//app.use("/stories", storyRoute);
-
+//app.use("/blog", blogRoutes);
+app.use("/newsletter", newsletterRoutes); 
+app.use("/stories", storyRoute);
 
 app.listen(PORT, () => {
     dbConnect();
     console.log(`Listening on PORT ${PORT}`);
   });
-
